third_party/blink/web_tests/svg/dynamic-updates/SVGFilterPrimitiveStandardAttributes-svgdom-result-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'result' propert of the SVGFilterPrimitiveStandardAttributes object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gradientElement.result.baseVal is "res1"
PASS gradientElement.result.baseVal is "res2"
PASS successfullyParsed is true

TEST COMPLETE
